+++ src/CleanRegex/Builder/PatternBuilder.php
@@ -4,7 +4,6 @@ namespace TRegx\CleanRegex\Builder;
 use TRegx\CleanRegex\Internal\Delimiter\Strategy\StandardStrategy;
 use TRegx\CleanRegex\Internal\Prepared\InjectParser;
 use TRegx\CleanRegex\Internal\Prepared\MaskParser;
-use TRegx\CleanRegex\Internal\Prepared\PreparedParser;
 use TRegx\CleanRegex\Internal\Prepared\PrepareFacade;
 use TRegx\CleanRegex\Internal\Prepared\Template\NoTemplate;
 use TRegx\CleanRegex\Pattern;
@@ -19,11 +18,6 @@ class PatternBuilder
     public function inject(string $input, array $values, string $flags = null): Pattern
     {
         return PrepareFacade::build(new InjectParser($input, $values, new NoTemplate()), new StandardStrategy($flags ?? ''));
-    }
-
-    public function prepare(array $input, string $flags = null): Pattern
-    {
-        return PrepareFacade::build(new PreparedParser($input), new StandardStrategy($flags ?? ''));
     }
 
     public function mask(string $mask, array $keywords, string $flags = null): Pattern