+++ test/Feature/CleanRegex/Builder/PatternBuilder/pcre/PatternBuilderTest.php
@@ -9,21 +9,6 @@ class PatternBuilderTest extends TestCase
     /**
      * @test
      */
-    public function shouldBuild_prepared()
-    {
-        // given
-        $pattern = Pattern::builder()->pcre()->prepare(['%You/her, (are|is) ', ['real? % (or are you not real?)'], ' (you|her)%']);
-
-        // when
-        $pattern = $pattern->delimited();
-
-        // then
-        $this->assertSame('%You/her, (are|is) real\?\ \%\ \(or\ are\ you\ not\ real\?\) (you|her)%', $pattern);
-    }
-
-    /**
-     * @test
-     */
     public function shouldBuild_inject()
     {
         // given